dev: Interim commit of SUDO logic
Most of the DEBUGGING is commented out.
The code compiles.
It mostly works.
However, there's a flaw in the logic...

Once it finds a strategy 1 match, it stops looking.
I need to refactor the logic to keep looking.
If a 2nd match is found, return ENODATA.
Otherwise, it found the real answer.

Changes to be committed:
	modified:   code/src/sudo_logic.c