dev: Initial commit of strategy two unit tests
I needed to be able to validate the strategy two functionality I'm
implementing without results being polluted by strategy one...

...so I wrote some strategy-specific unit tests, declared the
sudo_logic-internal function (so I could link against it), and run the
test cases.

SPOILERS: The row-based functionality seems to be working because all
but one test case is passing.

Changes to be committed:
	new file:   code/test/check_sudo_logic_solve_strategy_two.c

Changes not staged for commit:
	modified:   code/src/sudo_logic.c